nlp-python/requirements.txt
fastapi
uvicorn[standard]
langdetect
spacy
pydantic
python-multipart
httpx
google-generativeai
python-dotenv
